# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz API Documentation",
        "description": "API documentation for the Laravel application",
        "contact": {
            "email": "soycheanat2202@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/quiz/store": {
            "post": {
                "tags": [
                    "Quiz"
                ],
                "operationId": "31d830b662d9cb812a680e3cfe47cd1c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "type": {
                                                "type": "string"
                                            },
                                            "difficulty": {
                                                "type": "string"
                                            },
                                            "category": {
                                                "type": "string"
                                            },
                                            "question": {
                                                "type": "string"
                                            },
                                            "correct_answer": {
                                                "type": "string"
                                            },
                                            "incorrect_answers": {
                                                "type": "json"
                                            },
                                            "category_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "multiple",
                                    "difficulty": "medium",
                                    "category": "travelling",
                                    "question": "example content",
                                    "correct_answer": "example title",
                                    "incorrect_answers": "example content",
                                    "category_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "difficulty": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "category": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "question": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "correct_answer": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "incorrect_answer": {
                                            "type": "json",
                                            "example": "incorrect_answer=[]"
                                        },
                                        "category_id": {
                                            "type": "number",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quiz/{id}": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Get Detail Todo",
                "operationId": "ff37b75e9c424feb05ba11fc227a33ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "type"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "category"
                                        },
                                        "question": {
                                            "type": "string",
                                            "example": "question"
                                        },
                                        "correct_answer": {
                                            "type": "string",
                                            "example": "correct_answer"
                                        },
                                        "incorrect_answer": {
                                            "type": "json",
                                            "example": "incorrect_answer"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quiz/update/{id}": {
            "put": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Update Quiz",
                "operationId": "a78c23ce4a8f5e58d69726b9177447c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "type": {
                                                "type": "string"
                                            },
                                            "difficulty": {
                                                "type": "string"
                                            },
                                            "category": {
                                                "type": "string"
                                            },
                                            "question": {
                                                "type": "string"
                                            },
                                            "correct_answer": {
                                                "type": "string"
                                            },
                                            "incorrect_answers": {
                                                "type": "json"
                                            },
                                            "category_id": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "multiple",
                                    "difficulty": "medium",
                                    "category": "travelling",
                                    "question": "example content",
                                    "correct_answer": "example title",
                                    "incorrect_answers": "example content",
                                    "category_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "type"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "category"
                                        },
                                        "question": {
                                            "type": "string",
                                            "example": "question"
                                        },
                                        "correct_answer": {
                                            "type": "string",
                                            "example": "correct_answer"
                                        },
                                        "incorrect_answer": {
                                            "type": "json",
                                            "example": "incorrect_answer"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quiz/delete/{id}": {
            "delete": {
                "tags": [
                    "Quiz"
                ],
                "operationId": "03cc4765d977b40f820c6e79e0a260d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "quiz deleted success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}